requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
deepgram-sdk>=3.0.0
elevenlabs>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
httpx>=0.25.0
